requirements.txt
Flask==3.0.0
tensorflow==2.10.1
keras==2.10.0
h5py==3.9.0
numpy==1.25.2
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
Pillow==10.0.0
transformers==4.31.0
torch==2.0.1
tqdm==4.65.0
